+++ Samples~/2D Character Controller/Scripts/Character2DVisualizer.cs
@@ -1,5 +1,4 @@
-﻿using GameInput;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Samples.CharacterController2D.Scripts
 {
@@ -10,16 +9,19 @@ namespace Samples.CharacterController2D.Scripts
 
         //Unity Functions
         //============================================================================================================//
-        
+
+#if JAM_INPUT_DELEGATOR
         private void OnEnable()
         {
-            GameInputDelegator.OnMovementChanged += OnMovementChanged;
+            GameInput.GameInputDelegator.OnMovementChanged += OnMovementChanged;
         }
 
         private void OnDisable()
         {
-            GameInputDelegator.OnMovementChanged -= OnMovementChanged;
+            GameInput.GameInputDelegator.OnMovementChanged -= OnMovementChanged;
         }
+
+        
         //============================================================================================================//
         
         private void OnMovementChanged(Vector2 movementInput)
@@ -29,5 +31,19 @@ namespace Samples.CharacterController2D.Scripts
             
             spriteRenderer.flipX = movementInput.x < 0;
         }
+#else
+
+        private void LateUpdate()
+        {
+            var inputX = 0f;
+            Utilities.InputHelper.AxisInput(KeyCode.D, KeyCode.A, ref inputX);
+            
+            if (inputX == 0f)
+                return;
+            
+            spriteRenderer.flipX = inputX < 0;
+        }
+
+#endif
     }
 }