+++ CHANGELOG.md
@@ -36,10 +36,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - This includes `Editor Only` code that manually parses `FixedPaletteSettings` to obtain color values during the assembly compilation step
   - Added `UnityPaletteParser.cs` to parse & cache the `FixedPaletteSettings.asset` `yaml` file
     - This includes `YamlDotNet.dll`, which is used in-editor only to enable `UnityPaletteParser.cs`
+- Added **_Editor_** class `ScriptingDefinitionHelper.cs` to check for specific classes, adding or removing scripting defines
+  - Added check for `GameInputDelegator.cs`, and if it exists it will add `JAM_INPUT_DELEGATOR`
+- Added `InputHelper.cs` for function to process axis inputs using the Old Input system `KeyCode`
+- Added `AddPrefabIfExists.cs` as utility class that checks for a GUID within the Asset Database, adding it into the scene if it exists
+  - Added `--- GAME INPUT PROMISE ---.prefab` as container for the `GameInputDelegator` prefab GUID that will be added into the 2D & 3D Character controller sample scenes
 - Added `Geodesics Movement Examples` into Samples
   - Added shared assets such as Character with animations
   - Added `CharacterAnimationController.cs` into Shared assets to allow character animations
-  - Added `Input.cs` into `/SharedAssets/` to allow unified input processing for the examples
   - Added `/Sphere Examples/` sub-directory which contains the scene & `SphereMovementController.cs`
   - Added `/Torus Examples/` sub-directory which contains the scene & `TorusMovementController.cs`
 - Added `/Utilities/Geodesics/` directory for specialized maths
@@ -77,6 +81,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added optional parameter pitch to `SFXManager.PlaySound()`
 - Refactored `AddPackages.Packages` to merge the Package Ids & Package URLs into a single data field
 - Added nugetforunity package to `AddPackages.cs`
+- Added Checks for `JAM_INPUT_DELEGATOR` into 2D Character Controller Sample
+  - Added `#define OLD_INPUT_SYSTEM` into `CharacterController2D.cs` to allow for `[Conditional]` on new `ProcessInputs()` function
+  - Added `CharacterController2D.ProcessInputs()` as old input system fallback incase `GameInputDelegator.cs` doesn't exist
+  - Wrapped all calls to `GameInputDelegator.cs` in `CharacterController2D.cs` with `#if JAM_INPUT_DELEGATOR`
+  - Wrapped all calls to `GameInputDelegator.cs` in `Character2DVisualizer.cs` with `#if JAM_INPUT_DELEGATOR`
+  - Added `Character2DVisualizer.LateUpdate()` fallback if `!JAM_INPUT_DELEGATOR`
+- Added Checks for `JAM_INPUT_DELEGATOR` into 3D Character Controller Sample
+  - Added `#define OLD_INPUT_SYSTEM` into `CharacterController3D.cs` to allow for `[Conditional]` on new `ProcessInputs()` function
+  - Added `CharacterController3D.ProcessInputs()` as old input system fallback incase `GameInputDelegator.cs` doesn't exist
+- Changed `Geodesics Movement Examples` to use `InputHelper.cs` to process inputs
 
 ### Fixed
 -